fix: very slow on moderately large repository
This tool was taking 59-62 seconds to run on a repository with ~150,000 lines of code, on an M4 Mac. Profiling showed that the patsy.Cache wasn't caching; rather than figuring out what was going on there, I just used the paths and directories already present in the setup to make a map for resolving them, which took execution time down to about 3 seconds.

I'm not sure if there's something about my specific environment that is causing the cache to miss every time, but this change definitely makes the tool usable for me when it previously was not.

I'm not sure if it's OK to bump the Go version in go.mod. If it's important that it stay at 1.12 I suppose I could build it with that version.